+++ data/MOCK_DATA.csv
@@ -1,5 +1,5 @@
 id,name,brand,company,price,isbn
-1,Bumetanid,Bumetanide,Mylan Institutional Inc.,$42.28 ,158995707-5
+1,Bumetanide,Bumetanide,Mylan Institutional Inc.,$42.28 ,158995707-5
 2,"Juniperus virginiana, Lappa major, Rhamnus purshiana, Syzygium jambolanum, Chelidonium majus, Lycopodium clavatum, Myrtus communis, Natrum sulphuricum, Phosphoricum acidum, Phosphorus, Tarentula cuben",Pancreaforce,Apotheca Company,$21.18 ,042329202-1
 3,Aspirin,Low Dose Aspirin,"GREAT LAKES WHOLESALE, MARKETING, & SALES, INC.",$12.03 ,491425980-X
 4,Pulmo Ferrum,Pulmo Ferrumi,Uriel Pharmacy Inc.,$23.67 ,136397285-5
